server/requirements.txt
accelerate==0.31.0
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.136
botocore==1.34.136
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
clarifai==10.5.3
clarifai-grpc==10.5.4
click==8.1.7
cohere==5.5.8
coloredlogs==15.0.1
configobj==5.0.8
configparser==7.0.0
contextlib2==21.6.0
contourpy==1.2.1
crewai==0.41.1
crewai-tools==0.4.8
chroma-hnswlib==0.7.6
chromadb==0.5.5
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
docx2pdf==0.1.8
docx2txt==0.8
effdet==0.4.1
einops==0.8.0
email_validator==2.1.2
embedchain==0.1.120
embedchain-crewai==0.1.114
emoji==2.12.1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
filelock==3.15.1
filetype==1.2.0
flash-attn==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.0
fpdf2==2.7.9
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-aiplatform==1.57.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-cloud-vision==3.7.2
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
gptcache==0.1.43
GPUtil==1.4.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
inquirerpy==0.3.4
instructor==1.3.3
iopath==0.1.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.12
langchain-chroma==0.1.2
langchain-cohere==0.1.5
langchain-community==0.2.11
langchain-core==0.2.28
langchain-openai==0.1.7
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.82
layoutparser==0.3.4
looseversion==1.3.0
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
mem0ai==0.0.9
memzero==0.0.7
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
nodeenv==1.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.2.1
onnx==1.16.1
onnxruntime==1.18.0
openai==1.34.0
opencv-python==4.10.0.84
openpyxl==3.1.4
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.1
pfzy==0.3.4
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.16.0
pluggy==1.5.0
poppler-utils==0.1.0
portalocker==2.8.2
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
prov==2.0.1
psutil==6.0.0
pulsar-client==3.5.0
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyclipper==1.3.0.post5
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.0
pydantic_core==2.20.0
pydot==3.0.1
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.9.18
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
PyNaCl==1.5.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.372
pysbd==0.3.4
PySocks==1.7.1
pytesseract==0.3.10
pytest==8.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
python-rapidjson==1.18
pytube==15.0.0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
qdrant-client==1.10.1
rapidfuzz==3.9.3
rapidocr-onnxruntime==1.3.22
ratelimiter==1.2.0.post0
rdflib==6.3.2
regex==2023.12.25
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
schema==0.7.5
scikit-learn==1.5.1
scipy==1.13.1
selenium==4.22.0
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
shapely==2.0.4
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.4.1
tesseract==0.1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.15.2
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
traits==6.3.2
transformers==4.37.2
trio==0.26.0
trio-websocket==0.11.1
triton==2.3.1
tritonclient==2.47.0
typer==0.12.3
types-requests==2.32.0.20240602
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.6
unstructured-client==0.23.7
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.19.2


